backend/pyFlask/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary


# fastapi → Main FastAPI framework.
# uvicorn → ASGI server to run FastAPI.
# sqlalchemy → ORM for interacting with PostgreSQL.
# psycopg2-binary → PostgreSQL driver for Python.

